rename file terraform_resource_spec to external_resource_spec
