+++ reconcile/utils/external_resource_spec.py
