+++ reconcile/utils/terraform_client.py
@@ -14,7 +14,7 @@ from sretoolbox.utils import retry
 from sretoolbox.utils import threaded
 
 from reconcile.utils.aws_helper import get_region_from_availability_zone
-from reconcile.utils.terraform_resource_spec import ExternalResourceSpec, ExternalResourceSpecInventory
+from reconcile.utils.external_resource_spec import ExternalResourceSpec, ExternalResourceSpecInventory
 
 import reconcile.utils.lean_terraform_client as lean_tf
